refactor: userSettinginfo 훅 리팩토링 (#ATR-523)
