+++ apps/service/src/features/user-setting/api/fetchUserProfile.ts
@@ -1,0 +1,11 @@
+import { UserProfile } from '@/entities/profile'
+
+const fetchUserProfile = async (userId: string): Promise<UserProfile> => {
+  const data = await fetch(
+    `${process.env.NEXT_PUBLIC_API_URL}/api/v1/user/${userId}`,
+  ).then((res) => res.json())
+  const { user } = data
+  return user
+}
+
+export default fetchUserProfile
