+++ apps/service/src/features/user-setting/api/index.ts
@@ -1,2 +1,3 @@
+export { default as fetchUserProfile } from './fetchUserProfile'
 export { default as postDuplicateName } from './postDuplicateName'
 export { default as postUserSettingInfo } from './postUserSettingInfo'
