+++ apps/service/src/features/user-setting/lib/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useUserSetting } from './useUserSetting'
