+++ apps/service/src/features/user-setting/lib/index.ts
@@ -1,1 +1,2 @@
+export * from './hook'
 export * from './util'
